+++ lib/src/xml.ts
@@ -1,0 +1,10 @@
+/**
+ * A collection of tools for working with XML.
+ *
+ * @groupDescription XML
+ * @packageDocumentation
+ * @beta
+ */
+export { parseXml } from './xml/parseXml.js';
+export type { XmlNode } from './xml/XmlNode.js';
+export type { XmlParseOptions } from './xml/XmlParseOptions.js';